+++ packages/vrender-components/src/label/rect.ts
@@ -57,6 +57,21 @@ export class RectLabel extends LabelBase<RectLabelAttrs> {
       case 'inside-right':
         sx = 0.5;
         break;
+      case 'top-right':
+        sx = 0.5;
+        sy = -0.5;
+        break;
+      case 'top-left':
+        sx = -0.5;
+        sy = -0.5;
+        break;
+      case 'bottom-right':
+        sx = 0.5;
+        sy = 0.5;
+        break;
+      case 'bottom-left':
+        sx = -0.5;
+        sy = 0.5;
     }
 
     anchorX += sx * rectWidth;
@@ -75,6 +90,19 @@ export class RectLabel extends LabelBase<RectLabelAttrs> {
       vx = isInside ? 1 : -1;
     } else if (position.includes('right')) {
       vx = isInside ? -1 : 1;
+    }
+
+    switch (position) {
+      case 'top-right':
+      case 'bottom-right':
+        vx = -1;
+        break;
+      case 'top-left':
+      case 'bottom-left':
+        vx = 1;
+        break;
+      default:
+        break;
     }
 
     const x = anchorX + vx * offset + (vx * width) / 2;