+++ packages/vrender-components/src/marker/point.ts
@@ -24,6 +24,7 @@ import { IMarkPointItemPosition } from './type';
 import type { Point } from '../core/type';
 
 export class MarkPoint extends Marker<MarkPointAttrs> {
+  name = 'markPoint';
   static defaultAttributes = DEFAULT_MARK_POINT_THEME;
 
   private _item!: ISymbol | Tag | IImage | IRichText;
@@ -250,7 +251,8 @@ export class MarkPoint extends Marker<MarkPointAttrs> {
     };
 
     const line = new Segment({
-      points: []
+      points: [],
+      pickable: false // 组件容器本身不参与拾取
     });
     line.name = 'mark-point-line';
     this._line = line;
@@ -281,5 +283,9 @@ export class MarkPoint extends Marker<MarkPointAttrs> {
     this.setItemLineAttr(itemLine, position, itemPosition, itemLine?.visible);
     this.setDecorativeLineAttr(itemLine, itemPosition, itemLine?.decorativeLine?.visible);
     this.setItemAttributes(this._item, itemContent, itemPosition, type);
+  }
+
+  protected isValidPoints() {
+    return true;
   }
 }