+++ packages/vrender-core/src/core/stage.ts
@@ -1,5 +1,5 @@
 import type { IAABBBounds, IBounds, IBoundsLike, IMatrix } from '@visactor/vutils';
-import { AABBBounds, Bounds, Point } from '@visactor/vutils';
+import { AABBBounds, Bounds, Point, isString } from '@visactor/vutils';
 import type {
   IGraphic,
   IGroup,
@@ -235,7 +235,13 @@ export class Stage extends Group implements IStage {
 
     // 创建一个默认layer图层
     // this.appendChild(new Layer(this, this.global, this.window, { main: true }));
-    this.appendChild(this.layerService.createLayer(this, { main: true }));
+    this.appendChild(
+      this.layerService.createLayer(
+        this,
+        // 如果传入的canvas是string的id，就传入，避免被判定要使用离屏canvas
+        params.canvas && isString(params.canvas) ? { main: true, canvasId: params.canvas } : { main: true }
+      )
+    );
 
     this.nextFrameRenderLayerSet = new Set();
     this.willNextFrameRender = false;
@@ -290,6 +296,9 @@ export class Stage extends Group implements IStage {
     this.ticker.addTimeline(this.timeline);
     this.timeline.pause();
     this.optmize(params.optimize);
+    if (this._background && isString(this._background)) {
+      this.setAttributes({ background: this._background });
+    }
   }
 
   // 优化策略